+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.3.1]
+- Added `__tostring` to ResourceInterface
+- Added `__invoke` to ResourceLocatorInterface 
+- `ResourceLocationInterface::setPath` don't accept a `null` value anymore (produced an error anyway)
+- Added proper PHP7 type hints
+- Misc code quality and docblock fix 
+
 ## [4.3.0]
 - Dropping support for PHP 5.6 & 7.0
 - Updated rockettheme/toolbox to 1.4.x
@@ -39,6 +46,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Security
 -->
 
+[4.3.1]: https://github.com/userfrosting/uniformresourcelocator/compare/4.3.0...4.3.1
 [4.3.0]: https://github.com/userfrosting/uniformresourcelocator/compare/4.2.3...4.3.0
 [4.2.3]: https://github.com/userfrosting/uniformresourcelocator/compare/4.2.2...4.2.3
 [4.2.2]: https://github.com/userfrosting/uniformresourcelocator/compare/4.2.1...4.2.2